export message
