+++ v2/i18n/message_test.go
@@ -1,4 +1,4 @@
-package internal
+package i18n
 
 import (
 	"reflect"
