+++ v2/i18n/parse_test.go
@@ -1,4 +1,4 @@
-package internal
+package i18n
 
 import (
 	"reflect"
